split up pandas/tests/indexes/test_multi.py #18644
