+++ pandas/tests/indexes/multi/data/multiindex_v1.pickle
